feat : 캐러셀 UI 완성, 기능 추가 중
