+++ public/javascripts/api.js
@@ -1,0 +1,9 @@
+export function runAPI() {
+  let data;
+  const promise = fetch("http://localhost:3000/image")
+    .then(response => response.json())
+    .then(json => {
+      data = json;
+      console.log(data);
+    });
+}
