+++ public/stylesheets/main-top.scss
@@ -1,68 +1,162 @@
+@import "./carousel-first.scss";
 $outline_color: #cecfd1;
+$bg_color: #f3f3f3;
+$img-text_color: rgb(136, 136, 136);
+
 main {
-  width: 1276px;
+  width: 1274px;
   height: 660px;
   margin: 19px 600px;
   background-color: white;
   border: $outline_color 1px solid;
   justify-content: space-between;
   align-items: center;
+  border: $outline_color #cecfd1 1px solid;
 }
-.section-top {
-  outline: blue solid;
-  display: flex;
-  width: 100%;
-  height: 340px;
-}
-
 .section-top__container {
-  background-color: rosybrown;
-  height: 320px;
-  display: flex;
+  display: grid;
+  grid-template-rows: repeat(2, 340px); //340px높이 행 2개
+  grid-template-columns: repeat(5, 1fr); //1fraction의 비율로 5개의 칸
+  grid-auto-rows: 200px;
+}
+.item__img-box-wrap {
+  padding: 21px 27px;
 
-  #items-left {
-    position: flex;
-    outline: green solid;
-    height: 320px;
-    width: 510px;
-  }
-  #items-carousel {
-    position: flex;
-    display: inline-block;
-    outline: red solid;
-    height: 320px;
-    width: 763px;
+  .img-box {
+    width: 200px;
   }
 }
 
-.section-mid {
-  outline: yellowgreen solid;
-  position: flex;
+.section-top__item:nth-child(1) {
+  grid-column: 1 / 3; //1번 박스를 1번에서 3이 있는 위치까지 늘린다
+}
+
+.section-top__item:nth-child(2) {
+  grid-column: 3 / 6; //2번 박스를 3번에서 6이 있는 위치까지 늘린다
+}
+.item__carousel-wrap {
+  margin: 0 124.5px;
+  width: 514px;
+  height: 340px;
+  text-align: center;
+  //overflow: hidden;
+  position: relative;
+
+  .carousel-content {
+    position: absolute;
+    width: 514px;
+
+    /*가운데에서 시작*/
+    &:nth-child(1) {
+      transform: translate3d(0px, 0, 0);
+    }
+    &:nth-child(2) {
+      transform: translate3d(520px, 0, 0);
+    }
+    &:nth-child(3) {
+      transform: translate3d(-520px, 0, 0);
+    }
+  }
+}
+.panel-moving {
+  transition: tr;
+}
+
+.img-title {
   display: block;
-  width: 100%;
-  height: 200px;
+  margin-top: 10px;
+  margin-bottom: 1px;
+  font-size: 14px;
 }
-.main__mid {
-  .main__mid__items {
-    background-color: rosybrown;
-    height: 200px;
-    display: flex;
-    margin: 0px 10px;
+.img-text {
+  display: block;
+  font-size: 12px;
+  color: $img-text_color;
+  letter-spacing: -1px;
+  line-height: 14px;
+}
+.img-icon {
+  display: block;
+  float: left;
+  width: 31px;
+  height: 18px;
+  font-size: 12px;
+  background: url("../images/ico2.gif") no-repeat;
+  background-position: 0px, -850px;
+  text-indent: -9999px;
+}
 
-    li {
-      padding: 21px 27px;
-      margin: 0 10px;
-      height: 201px;
-      width: 276px;
-      background-color: wheat;
-      outline: black solid;
+.item__carousel-icon {
+  .inner-paging {
+    position: absolute;
+    left: 1470px;
+    top: 530px;
+
+    .btn-paging {
+      display: inline;
+      float: left;
+      margin-left: 3px;
+
+      .paging__num-selected {
+        background-color: #000;
+        display: block;
+        width: 15px;
+        height: 5px;
+        font-size: 0;
+        line-height: 0;
+        text-indent: -9999px;
+      }
+      .paging__num {
+        background-color: #cccccc;
+        display: block;
+        width: 15px;
+        height: 5px;
+        font-size: 0;
+        line-height: 0;
+        text-indent: -9999px;
+      }
+    }
+  }
+  #paging-btn-prev {
+    position: absolute;
+    left: 1168px;
+    top: 350px;
+    width: 63px;
+    height: 85px;
+    background-color: transparent;
+    border: none;
+
+    #prev {
+      display: inline-block;
+      width: 23px;
+      height: 45px;
+      background: url("../images/ico3.png") no-repeat;
+      background-position: -60px 0px;
+    }
+  }
+
+  #paging-btn-next {
+    position: absolute;
+    right: 660px;
+    top: 350px;
+    width: 63px;
+    height: 85px;
+    background-color: transparent;
+    border: none;
+
+    #next {
+      display: inline-block;
+      width: 23px;
+      height: 45px;
+      background: url("../images/ico3.png") no-repeat;
+      background-position: -150px 0px;
     }
   }
 }
 
-.main__view-more {
-  margin-top: 20px;
-  height: 100px;
-  width: 100%;
-  background-color: darkblue;
+.view-more {
+  height: 80px;
+  background-color: bisque;
+  text-align: center;
+  padding: 30px;
 }
