+++ public/stylesheets/style.scss
@@ -28,7 +28,6 @@ body {
 }
 
 header {
-  outline: red solid;
   display: block;
 }
