+++ views/index copy.html
@@ -55,120 +55,63 @@
 
     <main>
       <section class="section-top">
-        <div class="section-top__container">
-          <div class="section-top__item">
-            <img
-              src="http://shop4.daumcdn.net/shophow/sib/0_210219174900_cCDgZtbAhPiBQqcJamsODOWNoVrXILeM"
-            />
+        <article class="section-top__container">
+          <div class="main__top__items" id="items-left">
+            <a href="#" target ="_blank"class="link-evt">
+              <img src="http://shop4.daumcdn.net/shophow/sib/0_210219174900_cCDgZtbAhPiBQqcJamsODOWNoVrXILeM"/>
+            </a>
           </div>
-          <div class="section-top__item">
-            <div class="item__carousel-wrap">
-              <!-- 여기서 부터 캐러셀 -->
-              <div class="carousel-content" >
-                <a href="#"">
-                  <img
-                    src="//shop3.daumcdn.net/shophow/sib/0_210219175737_rZKhKlFtOPfWeMUVyowhqtRvzJJOintZ"
-                /></a>
-              </div>
-              <div class="carousel-content">
-                <a href="#">
-                  <img
-                    src="//shop2.daumcdn.net/shophow/sib/0_210219175824_lTVZxKLcBxDipjZVpYbrzginiMwRizAO"
-                  />
+          <div class="main__top__carousel" id="items-carousel">
+            <div class="top__carousel__evt">
+              <div class="slide">
+                <div class="panel"><a href="#" target ="_blank"class="link-evt">
+                  <img src="http://shop4.daumcdn.net/shophow/sib/0_210219174900_cCDgZtbAhPiBQqcJamsODOWNoVrXILeM"/>
                 </a>
               </div>
-              <div class="carousel-content">
-                <a href="#">
-                  <img
-                    src="//shop1.daumcdn.net/shophow/sib/0_210219175816_BLrjXAFKSXpUoiPzkxzQFtLjJkHixbZX"
-                  />
+                <div class="panel"><a href="#" target ="_blank"class="link-evt">
+                  <img src="http://shop4.daumcdn.net/shophow/sib/0_210219174900_cCDgZtbAhPiBQqcJamsODOWNoVrXILeM"/>
                 </a>
               </div>
-            </div>
+                <div class="panel"><a href="#" target ="_blank"class="link-evt">
+                  <img src="http://shop4.daumcdn.net/shophow/sib/0_210219174900_cCDgZtbAhPiBQqcJamsODOWNoVrXILeM"/>
+                </a>
+              </div>
+              </div>
 
-            <div class="item__carousel-icon">
-              <span class="inner-paging">
-                <span class="btn-paging">
-                  <span class="paging__num-selected">1</span>
+            </div>
+          </div>
+              <div class="items-carousel">
+                <span class="items-carousel__btn">
+                  <em class="btn__wrap"><span class="wrap__dot">0</span></em>
+                  <span class="btn__wrap"><span class="wrap__dot">1</span></span>
+                  <span class="btn__wrap"><span class="wrap__dot">2</span></span>
                 </span>
-                <span class="btn-paging"
-                  ><span class="paging__num">2</span></span
-                >
-                <span class="btn-paging"
-                  ><span class="paging__num">3</span></span
-                >
-              </span>
-              <button class="paging-btn" id="paging-btn-prev">
-                <span id="prev"></span>
-              </button>
-              <button class="paging-btn" id="paging-btn-next">
-                <span id="next"></span>
-              </button>
-            </div>
-          </div>
+                <button class="btn-prev" type="button" title="이전">
+                  <span class="btn-prev__icon">이전</span>
+                </button>
+                <button class="btn-next" type="button" title="다음">
+                  <span class="btn-next__icon">다음</span>
+                </button>
+              </div>
+        </article>
+      </section>
 
-          <div class="section-top__item">
-            <div class="item__img-box-wrap">
-              <img
-                class="img-box"
-                src="//shop1.daumcdn.net/thumb/S318x318/?fname=http%3A%2F%2Fshop2.daumcdn.net%2Fshophow%2Fp%2FO6937905659.jpg%3Fut%3D20190620152627&amp;scode=talkgift"
-              />
-              <strong class="img-title">라바트 가구 기획전</strong>
-              <span class="img-text">공간 전체를 아루르는 품격있는 분위기</span>
-              <span class="img-icon">테마</span>
-            </div>
-          </div>
-          <div class="section-top__item">
-            <div class="item__img-box-wrap">
-              <img
-                class="img-box"
-                src="//shop1.daumcdn.net/thumb/S318x318/?fname=http%3A%2F%2Fshop2.daumcdn.net%2Fshophow%2Fp%2FO6937905659.jpg%3Fut%3D20190620152627&amp;scode=talkgift"
-              />
-              <strong class="img-title">라바트 가구 기획전</strong>
-              <span class="img-text">공간 전체를 아루르는 품격있는 분위기</span>
-              <span class="img-icon">테마</span>
-            </div>
-          </div>
-          <div class="section-top__item">
-            <div class="item__img-box-wrap">
-              <img
-                class="img-box"
-                src="//shop1.daumcdn.net/thumb/S318x318/?fname=http%3A%2F%2Fshop2.daumcdn.net%2Fshophow%2Fp%2FO6937905659.jpg%3Fut%3D20190620152627&amp;scode=talkgift"
-              />
-              <strong class="img-title">라바트 가구 기획전</strong>
-              <span class="img-text">공간 전체를 아루르는 품격있는 분위기</span>
-              <span class="img-icon">테마</span>
-            </div>
-          </div>
-          <div class="section-top__item">
-            <div class="item__img-box-wrap">
-              <img
-                class="img-box"
-                src="//shop1.daumcdn.net/thumb/S318x318/?fname=http%3A%2F%2Fshop2.daumcdn.net%2Fshophow%2Fp%2FO6937905659.jpg%3Fut%3D20190620152627&amp;scode=talkgift"
-              />
-              <strong class="img-title">라바트 가구 기획전</strong>
-              <span class="img-text">공간 전체를 아루르는 품격있는 분위기</span>
-              <span class="img-icon">테마</span>
-            </div>
-          </div>
-          <div class="section-top__item">
-            <div class="item__img-box-wrap">
-              <img
-                class="img-box"
-                src="//shop1.daumcdn.net/thumb/S318x318/?fname=http%3A%2F%2Fshop2.daumcdn.net%2Fshophow%2Fp%2FO6937905659.jpg%3Fut%3D20190620152627&amp;scode=talkgift"
-              />
-              <strong class="img-title">라바트 가구 기획전</strong>
-              <span class="img-text">공간 전체를 아루르는 품격있는 분위기</span>
-              <span class="img-icon">테마</span>
-            </div>
-          </div>
+      <section class="section-mid">
+        <div class="main__mid">
+          <ul class="main__mid__items">
+            <li>
+              <div class="main__mid__img">아이템1</div>
+            </li>
+            <li><div class="main__mid__img">아이템1</div></li>
+            <li><div class="main__mid__img">아이템1</div></li>
+            <li><div class="main__mid__img">아이템1</div></li>
+            <li><div class="main__mid__img">아이템1</div></li>
+          </ul>
         </div>
-        <div class="view-more">
-          <span class="view-more___btn">더 보기(0 / 100 건) </span>
-        </div>
+        <div class="main__view-more">더 보기</div>
       </section>
     </main>
-    <!-- 
+
     <section class="carousel">
       <div class="carousel__title">지금뜨는 테마 카테고리</div>
       <ul class="carousel_items">
@@ -246,8 +189,8 @@
       최종 상품 정보는 구매 전 판매처에서 반드시 확인해주세요.<br>
       각 판매처의 매매보호 서비스를 통해 구매안전 절차 확인 후(에스크로/소비자피해보험/채무지금보증계약) 상품을 구매해 주시기 바랍니다.
     </p></footer>
-    <small class="copyright">Copyright © daisycommerce Corp. All rights reserved.</small> -->
+    <small class="copyright">Copyright © daisycommerce Corp. All rights reserved.</small>
 
-    <script type="module" src="../public/javascripts/main.js"></script>
+    <script src="../public/javascripts/main.js"></script>
   </body>
 </html>
